+++ PizzaShopAngular/src/app/components/pizza/pizza-card-cart-item/pizza-card-cart-item.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Cart} from '../../models/Cart';
 import {Pizza} from '../../models/Pizza';
 import {ThemeObjectService} from '../../../logic/theme-object/theme-object.service';
@@ -7,6 +7,8 @@ import {CartService} from '../../../logic/services/cartDao/cart.service';
 import {Drink} from '../../models/Drink';
 import {Snack} from '../../models/Snack';
 import {Dessert} from "../../models/Dessert";
+import {APiURL} from "../../../config/urlConfig";
+import {UserService} from "../../../logic/services/userDao/user.service";
 
 
 @Component({
@@ -22,15 +24,18 @@ export class PizzaCardCartItemComponent implements OnInit {
   @Input() snack: Snack;
   @Input() dessert: Dessert;
   @Input() pizzaPrice: number;
-  url = 'api/pizza/image/';
+  @Output() cartFromLs: EventEmitter<string> = new EventEmitter<string>();
+  url = APiURL.pizzaImage;
   count: number;
   price: number;
   cart: Cart;
-  urlDrink = '/api/drink/';
-  urlSnack = '/api/snack/';
-  urlDessert = '/api/dessert/';
+  urlDrink = APiURL.drinkImage;
+  urlSnack = APiURL.snackImage;
+  urlDessert = APiURL.dessertImage;
+
   constructor(private themeObjectService: ThemeObjectService,
-              private userService: UserActionsService,
+              private userService: UserService,
+              private userActionsService: UserActionsService,
               private cartService: CartService) {
   }
 
@@ -43,7 +48,6 @@ export class PizzaCardCartItemComponent implements OnInit {
 
   onInc(): void {
     try {
-      debugger;
       if (this.pizza) {
         this.cart = {
           id: this.cartElement.id,
@@ -53,7 +57,11 @@ export class PizzaCardCartItemComponent implements OnInit {
           amount: this.cartElement.amount + 1,
           size: this.cartElement.size
         };
-        this.cartService.addAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        if (this.userService.isAuthenticated()) {
+          this.cartService.addAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        } else if (!this.userService.isAuthenticated()) {
+          this.cart = this.cartService.onIncCartElementInLocalStorage(this.cart.description, this.pizzaPrice);
+        }
         this.price = this.price + this.pizzaPrice;
         this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice + this.pizzaPrice;
       } else if (this.drink) {
@@ -65,11 +73,14 @@ export class PizzaCardCartItemComponent implements OnInit {
           price: this.cartElement.price + this.drink.price,
           amount: this.cartElement.amount + 1,
         };
-        this.cartService.addAmountPizzaCart(this.cartElement.id, this.drink.price).subscribe(data => console.log(data));
+        if (this.userService.isAuthenticated()) {
+          this.cartService.addAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        } else if (!this.userService.isAuthenticated()) {
+          this.cart = this.cartService.onIncCartElementInLocalStorage(this.cart.description, this.pizzaPrice);
+        }
         this.price = this.price + this.drink.price;
         this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice + this.drink.price;
-      }
-      else if (this.snack) {
+      } else if (this.snack) {
         this.cart = {
           id: this.cartElement.id,
           description: this.cartElement.description,
@@ -77,11 +88,14 @@ export class PizzaCardCartItemComponent implements OnInit {
           price: this.cartElement.price + this.snack.price,
           amount: this.cartElement.amount + 1,
         };
-        this.cartService.addAmountPizzaCart(this.cartElement.id, this.snack.price).subscribe(data => console.log(data));
+        if (this.userService.isAuthenticated()) {
+          this.cartService.addAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        } else if (!this.userService.isAuthenticated()) {
+          this.cart = this.cartService.onIncCartElementInLocalStorage(this.cart.description, this.pizzaPrice);
+        }
         this.price = this.price + this.snack.price;
         this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice + this.snack.price;
-      }
-      else if (this.dessert) {
+      } else if (this.dessert) {
         this.cart = {
           id: this.cartElement.id,
           description: this.cartElement.description,
@@ -89,11 +103,17 @@ export class PizzaCardCartItemComponent implements OnInit {
           price: this.cartElement.price + this.dessert.price,
           amount: this.cartElement.amount + 1,
         };
-        this.cartService.addAmountPizzaCart(this.cartElement.id, this.dessert.price).subscribe(data => console.log(data));
+        if (this.userService.isAuthenticated()) {
+          this.cartService.addAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        } else if (!this.userService.isAuthenticated()) {
+          this.cart = this.cartService.onIncCartElementInLocalStorage(this.cart.description, this.pizzaPrice);
+        }
         this.price = this.price + this.dessert.price;
         this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice + this.dessert.price;
       }
-      this.userService.incAmountPizzaCartInStore(this.cartElement.id, this.cart);
+      if (this.userService.isAuthenticated()) {
+        this.userActionsService.incAmountPizzaCartInStore(this.cartElement.id, this.cart);
+      }
       this.count = this.count + 1;
     } catch (e) {
       console.log(e);
@@ -102,7 +122,6 @@ export class PizzaCardCartItemComponent implements OnInit {
 
   onDec(): void {
     try {
-      debugger;
       if (this.pizza) {
         this.cart = {
           id: this.cartElement.id,
@@ -112,7 +131,11 @@ export class PizzaCardCartItemComponent implements OnInit {
           amount: this.cartElement.amount - 1,
           size: this.cartElement.size
         };
-        this.cartService.removeAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        if (this.userService.isAuthenticated()) {
+          this.cartService.removeAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        } else if (!this.userService.isAuthenticated()) {
+          this.cartService.onDecCartElementInLocalStorage(this.cart.description, this.pizzaPrice);
+        }
         this.price = this.price - this.pizza.price;
         this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice - this.pizzaPrice;
       } else if (this.drink) {
@@ -124,11 +147,14 @@ export class PizzaCardCartItemComponent implements OnInit {
           price: this.cartElement.price - this.drink.price,
           amount: this.cartElement.amount - 1,
         };
-        this.cartService.removeAmountPizzaCart(this.cartElement.id, this.drink.price).subscribe(data => console.log(data));
+        if (this.userService.isAuthenticated()) {
+          this.cartService.removeAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        } else if (!this.userService.isAuthenticated()) {
+          this.cartService.onDecCartElementInLocalStorage(this.cart.description, this.pizzaPrice);
+        }
         this.price = this.price - this.drink.price;
         this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice - this.drink.price;
-      }
-      else if (this.snack) {
+      } else if (this.snack) {
         this.cart = {
           id: this.cartElement.id,
           description: this.cartElement.description,
@@ -136,11 +162,14 @@ export class PizzaCardCartItemComponent implements OnInit {
           price: this.cartElement.price - this.snack.price,
           amount: this.cartElement.amount - 1,
         };
-        this.cartService.removeAmountPizzaCart(this.cartElement.id, this.snack.price).subscribe(data => console.log(data));
+        if (this.userService.isAuthenticated()) {
+          this.cartService.removeAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        } else if (!this.userService.isAuthenticated()) {
+          this.cartService.onDecCartElementInLocalStorage(this.cart.description, this.pizzaPrice);
+        }
         this.price = this.price - this.snack.price;
         this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice - this.snack.price;
-      }
-      else if (this.dessert) {
+      } else if (this.dessert) {
         this.cart = {
           id: this.cartElement.id,
           description: this.cartElement.description,
@@ -148,20 +177,32 @@ export class PizzaCardCartItemComponent implements OnInit {
           price: this.cartElement.price - this.dessert.price,
           amount: this.cartElement.amount - 1,
         };
-        this.cartService.removeAmountPizzaCart(this.cartElement.id, this.dessert.price).subscribe(data => console.log(data));
+        if (this.userService.isAuthenticated()) {
+          this.cartService.removeAmountPizzaCart(this.cartElement.id, this.pizza.price).subscribe(data => console.log(data));
+        } else if (!this.userService.isAuthenticated()) {
+          this.cartService.onDecCartElementInLocalStorage(this.cart.description, this.pizzaPrice);
+        }
         this.price = this.price - this.dessert.price;
         this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice - this.dessert.price;
       }
-      this.userService.decAmountPizzaCartInStore(this.cartElement.id, this.cart);
+      if (this.userService.isAuthenticated()) {
+        this.userActionsService.decAmountPizzaCartInStore(this.cartElement.id, this.cart);
+      }
       this.count = this.count - 1;
     } catch (e) {
       console.log(e);
     }
   }
 
-  onDelete(id: number): void {
-    this.cartService.deleteCart(id).subscribe(data => console.log(data));
-    this.userService.deletePizzaCartInStore(id);
+  onDelete(id: number, description: string): void {
+    if (this.userService.isAuthenticated()) {
+      this.cartService.deleteCart(id).subscribe(data => console.log(data));
+      this.userActionsService.deletePizzaCartInStore(id);
+    } else {
+      this.cartService.deleteCartElementFromLocalStorage(description);
+      this.cartFromLs.emit(description);
+      this.themeObjectService.data.value.sizeCart--;
+    }
     this.themeObjectService.data.value.totalPrice = this.themeObjectService.data.value.totalPrice - this.cartElement.price;
   }
 }
